WIP: run persistenced as root
When persistenced is run as a non-root user it prints an error message
in the syslog saying that it was not able to open the device and then
the nvidia driver load hangs forever.

As a workaround this just run persistenced as root. Ideally we should
run it as non-root user as recommended in the documentation.