query_cc_mode: Check query_gpu_cc_mode() status
If the GPU is in CC mode, but the query_gpu_cc_mode() command fails to run,
the NVRC will silently proceed reporting that CC mode is off and refuse
to put the GPU in ready state.

This patch fixes that by checking the query_cc_mode command status and
returning an error if it failed.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>