Use nvidia-smi to set the GPU to ready state
The query_gpu_cc_mode() is exiting with error code 1 without printing
any log message. The reason is still unknown and it can't be reproduced
on every run.

This patch mitigates that by using "nvidia-smi conf-compute -f" to query
the GPU CC mode. With that the CC mode can be queried only after the
nvidia driver is loaded.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>